rough layout, private route, fake protected content
