+++ friends/src/App.js
@@ -1,26 +1,18 @@
 import React from 'react';
 import { Route, Link, Redirect } from 'react-router-dom'
+import PrivateRoute from './components/PrivateRoute'
+import Login from './components/Login'
+import Protected from './components/Protected'
 import './App.css';
 
 function App() {
 
-  const PrivateRoute = ({ component: Component, ...rest }) =>
-  {
-    <Route
-      {...rest}
-      render={props =>
-        localStorage.getItem("token") ? (
-          <Component {...props} />
-        ) : (
-          <Redirect to="/login" />
-        )
-      }
-    />
-  }
+  
 
   return (
     <div className="App">
-      
+      <Route path="/login" component={Login} />
+      <PrivateRoute path="/protected" component={Protected} />
     </div>
   );
 }
