+++ friends/src/components/Login.jsx
@@ -3,7 +3,7 @@ import React from "react"
 const Login = () =>
 {
     return (
-        <div></div>
+        <div>Login Screen</div>
     )
 }
