+++ friends/src/components/PrivateRoute.jsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import { Redirect, Route } from 'react-router-dom'
+
+const PrivateRoute = ({ component: Component, ...rest }) =>
+{
+    return(
+        <Route
+            {...rest}
+            render={props =>
+            localStorage.getItem("token") ? (
+                <Component {...props} />
+            ) : (
+                <Redirect to="/login" />
+            )
+            }
+        />
+    )
+}
+
+export default PrivateRoute
