+++ friends/src/components/Protected.jsx
@@ -1,0 +1,10 @@
+import React from "react"
+
+const Protected = () =>
+{
+    return (
+        <div>Protected Content</div>
+    )
+}
+
+export default Protected
